Refactor Progress and Final Riddle Pages to Use SidebarLayout Component
